+++ src/sxw.js
@@ -1,5 +1,5 @@
 /*
-* SXW JS
+* SXW.js
 * A simple javascript to print self XSS warning
 * GitHub    : https://github.com/sxw-js/sxw-js
 * Version   : v0.0.1 (2019-May-10)
@@ -9,23 +9,19 @@
 window.sxwjs = {
     config: {
         stopColor: "red",
-        supportTeam: "support team"
+        warningText: "\n\nThis section is intended for developers only. Don't copy paste anything in this area.\nIf someone told you to copy and paste something here, it is a scam and will give them access to your account. In that case, kindly report this to our support team."
     }
 };
 
-window.sxwjs.warningText = "\n\nThis section is intended for developers only. Don't copy paste anything in this area.\nIf someone told you to copy and paste something here, it is a scam and will give them access to your account. In that case, kindly report this to our " + window.sxwjs.config.supportTeam + ".";
-
 window.sxwjs.printStop = function () {
-    console.log("%c*********  *********  *********  *********\n***           ***     ***   ***  ***   ***\n***           ***     ***   ***  ***   ***\n*********     ***     ***   ***  *********\n      ***     ***     ***   ***  ***\n      ***     ***     ***   ***  ***\n*********     ***     *********  ***", "color:" + window.sxwjs.config.stopColor + "; font-weight:bold;");
+    console.log("%c*********  *********  *********  *********\n***           ***     ***   ***  ***   ***\n***           ***     ***   ***  ***   ***\n*********     ***     ***   ***  *********\n      ***     ***     ***   ***  ***\n      ***     ***     ***   ***  ***\n*********     ***     *********  ***", "color:" + window.sxwjs.config.stopColor + "; font-weight:900;");
 }
 
 window.sxwjs.printWarningText = function () {
-    console.log("%cDO NOT PROCEED.", "font-weight:bold; font-size:15px;", window.sxwjs.warningText);
+    console.log("%cCaution: DO NOT PROCEED.", "font-weight:bold; font-size:15px;", window.sxwjs.config.warningText);
 }
 
-window.sxwjs.printXssWarning = function printXssWarning() {
+window.sxwjs.printWarning = function () {
     window.sxwjs.printStop();
     window.sxwjs.printWarningText();
 }
-
-window.sxwjs.printXssWarning();